Remove pyanaconda.payload.livepayload
Move all code from livepayload.py to new files in pyanaconda.payload.live.
Rename the live payloads to LiveOSPayload and LiveImagePayload to reflect
the names of DBus modules.
