+++ configure.ac
@@ -141,6 +141,7 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/isys/Makefile
                  pyanaconda/payload/Makefile
                  pyanaconda/payload/dnf/Makefile
+                 pyanaconda/payload/live/Makefile
                  pyanaconda/payload/source/Makefile
                  pyanaconda/storage/Makefile
                  pyanaconda/ui/Makefile
